server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "backend of join0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.15",
    "fastapi>=0.116.1",
    "google-genai>=1.29.0",
    "groq>=0.31.0",
    "numpy>=2.3.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "requests>=2.32.0",
    "plotly>=5.22.0",
]
